+++ renderer/units/JavaScript Runner/samples/CJK/Unzip CJK Sources Data Files to User Data Directory.js
@@ -1,4 +1,4 @@
-// Unzip SVG Glyphs Data Files to User Data Directory
+// Unzip CJK Sources Data Files to User Data Directory
 const path = require ('path');
 const extract = require ('extract-zip');
 // Note: the relevant zip file is assumed to be in the Downloads directory